Init groups on navbar mount
This allows the group colors to be displayed even if the Group Filter is initially hidden.
